+++ src/main/java/com/jalasoft/sdfc/pages/AbstractPage.java
@@ -6,7 +6,7 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-public class AbstractPage {
+abstract class AbstractPage {
 
 	protected WebDriver driver;
 	protected WebDriverWait wait;
